# crates/server/openapi.yaml
...
openapi: 3.0.3
info:
  title: Remote Harness Control Plane API
  version: 0.1.0
  description: |
    Control plane REST API. Human-readable API prose and behavior notes live in
    `docs/API_OVERVIEW.md`; this spec is the machine-readable contract for codegen and review.
servers:
  - url: /
    description: Relative to configured control plane base URL
...
paths:
...
  /sessions:
    get:
      tags: [sessions]
      summary: List sessions
      operationId: listSessions
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            description: Filter by session status (e.g. pending, running, completed, failed)
      responses:
        "200":
          description: Paginated sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSessionSummaries"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
    post:
      tags: [sessions]
      summary: Create session (start workflow)
      operationId: createSession
      security:
        - bearerApiKey: []
        - headerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
  /identities/{id}:
    get:
      tags: [identities]
      summary: Get identity credential flags (no secret values)
      operationId: getIdentity
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: Identity id (e.g. default)
      responses:
        "200":
          description: Whether git and agent tokens are stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityCredentials"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Identity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
    patch:
      tags: [identities]
      summary: Update identity tokens (merge; secrets never returned)
      operationId: patchIdentity
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchIdentityRequest"
      responses:
        "204":
          description: Updated
        "400":
          description: Invalid body or empty patch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Identity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
  /identities/{id}/auth-status:
    get:
      tags: [identities]
      summary: Token health for Git (and agent presence) without secret values
      operationId: getIdentityAuthStatus
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Auth status summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityAuthStatus"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Identity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
  /identities/{id}/repositories:
    get:
      tags: [identities]
      summary: List repositories visible to the stored Git token
      operationId: listIdentityRepositories
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: provider
          in: query
          required: false
          schema:
            type: string
            enum: [github, gitlab]
          description: Required for manual PAT identities; implied for OAuth identities
      responses:
        "200":
          description: Repo picker items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityRepositoriesResponse"
        "400":
          description: Missing provider or no git token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing API key or Git provider rejected token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Identity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "502":
          description: Git provider API or network error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      description: "API key as Bearer token (`Authorization: Bearer <api_key>`)"
    headerApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key in `X-API-Key` header"
  schemas:
...
    StandardError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Resource not found
            details:
              type: object
              additionalProperties: true
    WorkflowKind:
      type: string
      enum:
        - chat
        - loop_n
        - loop_until_sentinel
        - inbox
    CreateSessionRequest:
      type: object
      required: [repo_url, workflow, params]
      properties:
        repo_url:
          type: string
        ref:
          type: string
          description: Git ref (branch or commit); default main when omitted
        workflow:
          $ref: "#/components/schemas/WorkflowKind"
        params:
          type: object
          additionalProperties: true
          description: Workflow-specific parameters (see API_OVERVIEW)
        persona_id:
          type: string
        identity_id:
          type: string
        retain_forever:
          type: boolean
          default: false
    CreateSessionResponse:
      type: object
      required: [session_id, status]
      properties:
        session_id:
          type: string
        status:
          type: string
          description: pending until first job assign (chat/loop); running for inbox with no initial jobs
          example: pending
        web_url:
          type: string
          description: Optional deep link to Web UI for this session
    SessionSummary:
      type: object
      required:
        - session_id
        - repo_url
        - ref
        - workflow
        - status
        - created_at
      properties:
        session_id:
          type: string
        repo_url:
          type: string
        ref:
          type: string
        workflow:
          type: string
        status:
          type: string
        created_at:
          type: string
          description: ISO 8601 timestamp
    PaginatedSessionSummaries:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
        next_cursor:
          type: string
          nullable: true
...
    IdentityCredentials:
      type: object
      required: [has_git_token, has_agent_token]
      properties:
        has_git_token:
          type: boolean
        has_agent_token:
          type: boolean
    GitTokenStatus:
      type: string
      enum:
        - healthy
        - expiring_soon
        - expired_refreshable
        - expired_needs_reauth
        - unknown
        - not_configured
    IdentityAuthStatus:
      type: object
      required: [git_token_status, message]
      properties:
        git_token_status:
          $ref: "#/components/schemas/GitTokenStatus"
        git_provider:
          type: string
          nullable: true
          description: manual | oauth_github | oauth_gitlab when known
        token_expires_at:
          type: string
          nullable: true
          description: ISO 8601 when expiry is tracked
        message:
          type: string
        agent_token_status:
          type: string
          nullable: true
          description: not_configured | healthy when agent token presence is summarized
    IdentityRepositoryItem:
      type: object
      required: [full_name, clone_url]
      properties:
        full_name:
          type: string
        clone_url:
          type: string
    IdentityRepositoriesResponse:
      type: object
      required: [items, provider]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IdentityRepositoryItem"
        provider:
          type: string
          enum: [github, gitlab]
    PatchIdentityRequest:
      type: object
      description: |
        JSON merge patch for stored tokens. Omit a property to leave it unchanged.
        Set a property to JSON null to clear stored ciphertext. Plain strings set new values.
      properties:
        agent_token:
          type: string
          nullable: true
        git_token:
          type: string
          nullable: true
        refresh_token:
          type: string
          nullable: true